REF: Make checksum comparison optional
To determine if a file has changed we were using the checksum in the newer versions, but eslint was relying on the old behavior where we use the mtime and file size to determine if a file changed. That's why we decided to make the checksum check optional.

To use it:

```js
// to make the cache use the checkSum check do the following:
var fCache = fileEntryCache.create(cacheName, dir, useCheckSum); // pass the third parameter as true
var otherCache = fileEntryCache.createFromFile(cacheName, useCheckSum); // pass the second parameter as true
```